ajuste de delete

+++ Api/src/service/index.js
@@ -25,7 +25,11 @@ class GatoService{
   }
 
   async deletar(id){
-    await this.gato.destroy({ where: { id: id } })
+    if (await this.gato.findOne({where: { id: id}})){
+      return await this.gato.destroy({ where: { id: id } })
+    }else{
+      return false
+    }
   }
 }
